Fallback for broken images. Closes #62
Note that we are doing a small optimization for the left hand-side avatar image (conditional new element creation depending on `onerror` function). If the image failed to load in the past then the `onerror` is already set to `null` and it is not enough to re-set it again because some browsers (like Chrome) will execute the `onerror` only once, thus we have to create a new element and replace the old img element.